hush-observability/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hush-observability"
version = "0.1.0"
description = "Observability and tracing framework for Hush with support for multiple backends"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Hush Team" }
]
dependencies = [
    "hush-core>=0.1.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
langfuse = [
    "langfuse>=2.6.0,<3.0.0",
]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]
all = [
    "langfuse>=2.6.0,<3.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["hush*"]
namespaces = true

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(levelname)s %(name)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests (may require external services)",
]

[tool.uv.sources]
hush-core = { path = "../hush-core", editable = true }
hush-providers = { path = "../hush-providers", editable = true }

[dependency-groups]
dev = [
    "langfuse>=2.6.0,<3.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.0",
]
